requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy
pyodbc
google-generativeai